Refactor each command into their own file.
